+++ little_brother/test/test_status_server_devices.py
@@ -21,7 +21,11 @@
 import unittest
 
 from little_brother import constants
+from little_brother import dependency_injection
 from little_brother import status_server
+from little_brother.persistence.persistent_device import Device
+from little_brother.persistence.persistent_device_entity_manager import DeviceEntityManager
+from little_brother.persistence.session_context import SessionContext
 from little_brother.test.base_test_status_server import BaseTestStatusServer
 from python_base_app.test import base_test
 
@@ -30,31 +34,59 @@ class TestStatusServerDevices(BaseTestStatusServer):
 
     @base_test.skip_if_env("NO_SELENIUM_TESTS")
     def test_page_devices(self):
-        _status_server = self.create_dummy_status_server()
+        self.create_dummy_status_server()
+        self.create_selenium_driver()
+
         self._status_server.start_server()
 
-        driver = self.create_selenium_driver()
+        self.login_devices()
 
-        # When we load the admin page the first time...
-        self._driver.get(self._status_server.get_url(p_internal=False, p_rel_url=status_server.DEVICES_REL_URL))
-        assert constants.APPLICATION_NAME in self._driver.title
-
-        # ...we end up on the login page.
-        self.login()
-
-        # After logging in we are on the users page
-        xpath = "//FORM/DIV/DIV[DIV[1] = 'Device' and DIV[2] = '']"
-        assert "Device Configuration" in self._driver.title
-        self._driver.find_element_by_xpath(xpath)
-
-        # The second time we call the users page.
+        # The second time we call the device page.
         self._driver.get(self._status_server.get_url(p_internal=False, p_rel_url=status_server.DEVICES_REL_URL))
         assert constants.APPLICATION_NAME in self._driver.title
         assert "Device Configuration" in self._driver.title
 
         # we are on the users page right away...
-        xpath = "//FORM/DIV/DIV[DIV[1] = 'Device' and DIV[2] = '']"
+        self.check_empty_device_list()
+
+    @base_test.skip_if_env("NO_SELENIUM_TESTS")
+    def test_page_devices_add_and_delete_device(self):
+        self.create_dummy_status_server()
+        self.create_selenium_driver()
+
+        self._status_server.start_server()
+
+        self.login_devices()
+
+        device_entity_manager: DeviceEntityManager = dependency_injection.container[DeviceEntityManager]
+
+        with SessionContext(self._persistence) as session_context:
+            new_device_name = device_entity_manager.get_new_device_name(
+                p_session_context=session_context, p_name_pattern=constants.DEFAULT_DEVICE_NEW_NAME_PATTERN)
+
+            add_button = self._driver.find_element_by_id("add_device")
+            add_button.click()
+
+            device: Device = device_entity_manager.get_by_device_name(
+                p_session_context=session_context, p_device_name=new_device_name)
+            self.assertEqual(new_device_name, device.device_name)
+
+        device_id = device.id
+
+        xpath = "//DIV/A[@aria-controls='detailsdevice_1']"
         self._driver.find_element_by_xpath(xpath)
+
+        delete_button = self._driver.find_element_by_id("delete_device_1")
+        delete_button.click()
+
+        delete_button = self._driver.find_element_by_id("delete_device_1-modal-confirm")
+        # See https://stackoverflow.com/questions/56194094/how-to-fix-this-issue-element-not-interactable-selenium-python
+        self._driver.execute_script("arguments[0].click();", delete_button)
+
+        with SessionContext(self._persistence) as session_context:
+            device = device_entity_manager.get_by_id(
+                p_session_context=session_context, p_id=device_id)
+            self.assertIsNone(device)
 
 
 if __name__ == "__main__":